+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 <!--@eqol-beta@-->
 
+## [11.4.0-beta2] - 2026-06-17
+
+### ✨ Added
+
+- Mythic Plus / Bloodlust Tracker: Added optional Bloodlust duration and active glow display while the lockout is freshly applied.
+- Minimap / Button Sink: Added an option to hide the Button Sink toggle itself while still collecting minimap buttons.
+
+---
+
 ## [11.4.0-beta1] - 2026-06-17
 
 ### ✨ Added